Invert the layering so that rest server feature is added last so we can reuse lower layer for non-REST uses.  Disable LoadBalancerTests - most are defunct.
