+++ core/loadBalancer/src/whisk/core/loadBalancer/LoadBalancerService.scala
@@ -16,6 +16,7 @@
 
 package whisk.core.loadBalancer
 
+import akka.actor.Actor
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
 import scala.util.Failure
@@ -25,53 +26,61 @@ import spray.http.StatusCodes.InternalServerError
 import spray.http.StatusCodes.OK
 import spray.httpx.SprayJsonSupport._
 import spray.json.DefaultJsonProtocol._
+import spray.json.JsBoolean
 import spray.json.JsObject
+import whisk.common.ConsulKV
+import whisk.common.ConsulKV.LoadBalancerKeys
+import whisk.common.ConsulKVReporter
 import whisk.common.Logging
 import whisk.common.TransactionId
+import whisk.common.Verbosity
+import whisk.core.WhiskConfig
 import whisk.core.connector.{ ActivationMessage => Message }
 import whisk.core.connector.LoadBalancerResponse
+import whisk.connector.kafka.KafkaProducerConnector
 import whisk.http.BasicRasService
+import whisk.utils.ExecutionContextFactory
 
-trait LoadBalancerService
-    extends BasicRasService
+class LoadBalancerService(config: WhiskConfig, verbosity: Verbosity.Level)
+    extends LoadBalancerToKafka
     with Logging {
 
     /** The execution context for futures */
-    implicit val executionContext: ExecutionContext
+    implicit val executionContext = ExecutionContextFactory.makeExecutionContext()
 
-    override def routes(implicit transid: TransactionId) = super.routes ~ publish ~ invokers
+    /**
+     * The two public methods are getInvokerHealth and the inherited doPublish methods.
+     */
+    def getInvokerHealth(): JsObject = invokerHealth.getInvokerHealth()
 
-    def publish(implicit transid: TransactionId) = {
-        (path("publish" / s"""(${Message.ACTIVATOR}|${Message.INVOKER})""".r) & post & entity(as[Message])) {
-            (component, message) =>
-                onComplete(doPublish(component, message)(message.transid)) {
-                    case Success(response) => complete(OK, response)
-                    case Failure(t)        => complete(InternalServerError)
-                }
-        }
-    }
+    override def getInvoker(message: Message): Option[Int] = invokerHealth.getInvoker(message)
+    override def activationThrottle = _activationThrottle
 
-    val invokers = {
-        (path("invokers") & get) {
-            complete {
-                getInvokerHealth()
+    override val producer = new KafkaProducerConnector(config.kafkaHost, executionContext)
+
+    private val kvStore = new ConsulKV(config.consulServer)
+    private val invokerHealth = new InvokerHealth(config, { () => producer.sentCount() })
+    private val _activationThrottle = new ActivationThrottle(LoadBalancer.config.consulServer, invokerHealth)
+
+    // --- WIP -----
+    private var count = 0
+    private val overloadThreshold = 5000 // this is the total across all invokers.  Disable by setting to -1.
+    private val reporter = new ConsulKVReporter(kvStore, 3000, 2000,
+        LoadBalancerKeys.hostnameKey,
+        LoadBalancerKeys.startKey,
+        LoadBalancerKeys.statusKey,
+        { () =>
+            val producedCount = getActivationCount()
+            val consumedCounts = invokerHealth.getInvokerActivationCounts()
+            val inFlight = producedCount - consumedCounts.fold(0) { (total, n) => total + n }
+            val overload = JsBoolean(overloadThreshold > 0 && inFlight >= overloadThreshold)
+            count = count + 1
+            if (count % 10 == 0) {
+                warn(this, s"In flight: ${producedCount} - [${consumedCounts.mkString(", ")}] = ${inFlight} ${overload}")
             }
-        }
-    }
+            Map(LoadBalancerKeys.overloadKey -> overload,
+                LoadBalancerKeys.invokerHealth -> getInvokerHealth()) ++
+                getUserActivationCounts()
+        })
 
-    /**
-     * Handles POST /publish/topic URI.
-     *
-     * @param component the component name extracted from URI (invoker, or activator)
-     * @param msg the Message received via POST
-     * @return response to terminate HTTP connection with
-     */
-    def doPublish(component: String, msg: Message)(implicit transid: TransactionId): Future[LoadBalancerResponse]
-
-    /**
-     * Handles GET /invokers URI.
-     *
-     * @return JSON of invoker health
-     */
-    def getInvokerHealth(): JsObject
 }
